requirements.txt
asgiref==3.8.1
chardet==5.2.0
click==8.1.7
Django==5.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
pillow==11.1.0
platformdirs==4.2.2
PyJWT==2.9.0
PyPDF2==3.0.1
reportlab==4.2.5
sqlparse==0.5.1
